+++ src/main/java/ru/native_speakers/cinema_expert_api/repository/ImagesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.native_speakers.cinema_expert_api.model.Image;
 
 @Repository
-public interface ImagesRepository extends JpaRepository<Image, Integer> {
+public interface ImagesRepository extends JpaRepository<Image, Long> {
 }
